🐛 fix(be): Add macOS development compatibility support
- Add darwin build constraints to CGo metadata
- Split osutil/volume_service per-platform
- Add darwin vendor stubs for mount/loop/sysfs
- Initialize default network stats on construction
- Skip tests requiring Linux-only env (ntfsfix, proc)
- Hook darwin stub regeneration into mise patch task
