+++ backend/.mise.toml
@@ -327,7 +327,11 @@ run = """
 		echo "ERROR: 'patch' command not found. Please install patch utility."; 
 		exit 1; 
 	fi; 
-	go -C src mod vendor; 
+	go -C src mod vendor;
+	# Regenerate Darwin vendor stubs (deleted by go mod vendor) for macOS dev
+	if [ "$(uname -s)" = "Darwin" ]; then
+		sh patches/create_darwin_stubs.sh src;
+	fi
 	find src/vendor -name "*.rej" -delete 2>/dev/null
 	#if [ -f "patches/goenums.patch" ] && [ -d "src/vendor/github.com/zarldev/goenums" ]; then 
 	#	echo "Applying goenums.patch..."; 
@@ -385,7 +389,7 @@ flag "--version [version]" help="Build version"
 run =[
 	'echo "Generating metadata constants ${usage_version:?}..."',
   'echo "// Code generated by mise task; DO NOT EDIT." > src/config/metadata_constants.go',
-	'echo "//go:build !cgo" >> src/config/metadata_constants.go',
+	'echo "//go:build !cgo || !linux" >> src/config/metadata_constants.go',
 	'echo "" >> src/config/metadata_constants.go',
 	'echo "package config" >> src/config/metadata_constants.go',
 	'echo "const MetadataJSON = \"SRAT_METADATA:{\\\"version\\\": \\\"v${usage_version:?}\\\"}\"" >> src/config/metadata_constants.go',
