+++ backend/src/config/c_medatata.go
@@ -1,4 +1,4 @@
-//go:build cgo
+//go:build linux && cgo
 
 package config
