+++ backend/src/dto/filesystem_mount_flags_darwin.go
@@ -1,0 +1,7 @@
+//go:build darwin
+
+package dto
+
+func MountFlagsMap() map[string]uintptr {
+	return map[string]uintptr{}
+}
