+++ backend/src/dto/filesystem_mount_flags_linux.go
@@ -1,0 +1,31 @@
+//go:build linux
+
+package dto
+
+import "syscall"
+
+func MountFlagsMap() map[string]uintptr {
+	return map[string]uintptr{
+		"ro":          syscall.MS_RDONLY,
+		"nosuid":      syscall.MS_NOSUID,
+		"nodev":       syscall.MS_NODEV,
+		"noexec":      syscall.MS_NOEXEC,
+		"sync":        syscall.MS_SYNCHRONOUS,
+		"remount":     syscall.MS_REMOUNT,
+		"mand":        syscall.MS_MANDLOCK,
+		"dirsync":     syscall.MS_DIRSYNC,
+		"noatime":     syscall.MS_NOATIME,
+		"nodiratime":  syscall.MS_NODIRATIME,
+		"bind":        syscall.MS_BIND,
+		"rec":         syscall.MS_REC,
+		"silent":      syscall.MS_SILENT,
+		"posixacl":    syscall.MS_POSIXACL,
+		"acl":         syscall.MS_POSIXACL,
+		"unbindable":  syscall.MS_UNBINDABLE,
+		"private":     syscall.MS_PRIVATE,
+		"slave":       syscall.MS_SLAVE,
+		"shared":      syscall.MS_SHARED,
+		"relatime":    syscall.MS_RELATIME,
+		"strictatime": syscall.MS_STRICTATIME,
+	}
+}
