+++ backend/src/dto/filesystem_type_test.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package dto
 
 import (
