+++ backend/src/service/filesystem_service_test.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package service_test
 
 import (
