+++ backend/src/service/volume_service_udev_darwin.go
@@ -1,0 +1,11 @@
+//go:build darwin
+
+package service
+
+import (
+	"log/slog"
+)
+
+func (self *VolumeService) udevEventHandler() {
+	slog.WarnContext(self.ctx, "Udev event handler is not supported on this platform")
+}
